mofification page d'acceuil et page produit

+++ apps/front/src/app/page/tab1/tab1.page.ts
@@ -1,17 +1,24 @@
 import { Component, OnInit } from '@angular/core';
 import { WebApiService } from '../../service/web-api.service';
-
+import { ActionSheetController } from '@ionic/angular';
 @Component({
   selector: 'app-tab1',
   templateUrl: 'tab1.page.html',
   styleUrls: ['tab1.page.scss']
 })
 export class Tab1Page implements OnInit {
+openActionSheet(arg0: any) {
+throw new Error('Method not implemented.');
+}
   storages: any;
   isEditing: boolean = false;
   selectedStorageId: number | null = null;
 
-  constructor(private webApiService: WebApiService) {}
+  constructor(
+    private webApiService: WebApiService,
+    public actionSheetController: ActionSheetController
+  ) {}
+
 
   ngOnInit() {
     this.getStorages();
@@ -24,7 +31,11 @@ export class Tab1Page implements OnInit {
     });
   }
 
-  deleteStorage(id: number) {
+  deleteStorage(id: number | null) {
+    if (id === null) {
+      console.error('No storage selected');
+      return;
+    }
     this.webApiService.deleteStorage(id).subscribe(() => {
       console.log('Stockage supprimé avec succès.');
       this.getStorages();
@@ -49,15 +60,52 @@ export class Tab1Page implements OnInit {
     }
   }
 
-
-  enableEdit(id: number) {
+  enableEdit(id: number | null) {
+    if (id === null) {
+      console.error('No storage selected');
+      return;
+    }
     this.isEditing = true;
     this.selectedStorageId = id;
   }
 
-  cancelEdit() {
-    this.isEditing = false;
-    this.selectedStorageId = null;
+  async actionSheetButtons(id: number) {
+    this.selectedStorageId = id;
+    const actionSheet = await this.actionSheetController.create({
+      header: 'Actions',
+      buttons: [
+        {
+          text: 'Delete',
+          role: 'destructive',
+          handler: () => {
+            this.deleteStorage(this.selectedStorageId);
+          },
+        },
+        {
+          text: 'Edit',
+          handler: () => {
+            this.enableEdit(this.selectedStorageId);
+          },
+        },
+        {
+          text: 'Cancel',
+          role: 'cancel',
+          handler: () => {
+            console.log('Cancel clicked');
+          },
+        },
+      ],
+    });
+    await actionSheet.present();
   }
-  
+
+  createNewStorage(nom: string) {
+    this.webApiService.createStorage(nom).subscribe(() => {
+        console.log('Stockage créé avec succès.');
+        this.getStorages();
+    }, erreur => {
+        console.log('Erreur lors de la création du stockage :', erreur);
+    });
+}
+
 }
